+++ Robust.Shared/Toolshed/ToolshedCommandImplementor.cs
@@ -84,7 +84,6 @@ internal sealed class ToolshedCommandImplementor
             ctx.Error.Contextualize(ctx.Input, (ctx.Bundle.NameStart, ctx.Bundle.NameEnd));
             return false;
         }
-        ctx.Bundle.CommandMethod = new CurrentCommandMethod(method.Value.Info, method.Value.Args);
 
         var argsStart = ctx.Index;
         if (!TryParseArguments(ctx, method.Value))
@@ -110,6 +109,7 @@ internal sealed class ToolshedCommandImplementor
 
         foreach (var arg in method.Args)
         {
+            ctx.CurrentArgument = arg;
             object? parsed;
             if (arg.IsParamsCollection)
             {
@@ -124,6 +124,7 @@ internal sealed class ToolshedCommandImplementor
             ctx.Bundle.Arguments[arg.Name] = parsed;
         }
 
+        ctx.CurrentArgument = null;
         DebugTools.AssertNull(ctx.Error);
         DebugTools.AssertNull(ctx.Completions);
         return true;
@@ -156,7 +157,6 @@ internal sealed class ToolshedCommandImplementor
     {
         DebugTools.AssertNull(ctx.Error);
         DebugTools.AssertNull(ctx.Completions);
-        ctx.CurrentArgument = arg;
         var start = ctx.Index;
         var save = ctx.Save();
         ctx.ConsumeWhitespace();
@@ -378,13 +378,18 @@ internal sealed class ToolshedCommandImplementor
             argType = argType.GetElementType()!;
         }
 
+        // TODO TOOLSHED
+        // Find a good way to give parsers access to argument attributes.
+        // I don't want to give just content access to all (possibly internal) attributes/
+        // Currently ListLengthAttribute is just hardcoded.
         return new CommandArgument(
             arg.Name!,
             argType,
             GetArgumentParser(arg, argType),
             arg.IsOptional,
             arg.DefaultValue,
-            isParamsCollection);
+            isParamsCollection,
+            arg.GetCustomAttribute<ListLengthAttribute>());
     }
 
     private ITypeParser? GetArgumentParser(ParameterInfo param, Type type)
@@ -765,9 +770,8 @@ public readonly record struct CommandArgument(
     ITypeParser? Parser,
     bool IsOptional,
     object? DefaultValue,
-    bool IsParamsCollection);
-
-public readonly record struct CurrentCommandMethod(MethodInfo Info, CommandArgument[] Args);
+    bool IsParamsCollection,
+    ListLengthAttribute? ListLengthAttribute);
 
 public sealed class ArgumentParseError(Type type, Type parser) : ConError
 {